fixed the program creation error where the program list would not update after program creation
