+++ src/components/program/modals/EditConfirm.js
@@ -53,13 +53,29 @@ const EditConfirm = (props) => {
 
         serverHandshake(true).put("/programs", updatedProgram)
         .then(res => {
+            serverHandshake(true).get('/programs')
+            .then(res => {
             Dispatch({ type: 'SET_PROGRAM_DATA', payload: res.data });
             Dispatch({ type: "UPDATE_NEW_PROGRAM_DATA", payload: defaultProgram });
+            })
+            .catch(err => {
+            console.log("there was an error", err);
+            })
             props.toggleConfirmModal(false);
             props.navigate("/program");
         })
         .catch(err => {
             console.log("there was an error", err);
+            serverHandshake(true).get('/programs')
+            .then(res => {
+            Dispatch({ type: 'SET_PROGRAM_DATA', payload: res.data });
+            Dispatch({ type: "UPDATE_NEW_PROGRAM_DATA", payload: defaultProgram });
+            })
+            .catch(err => {
+            console.log("there was an error", err);
+            })
+            props.toggleConfirmModal(false);
+            props.navigate("/program");
         })
     }
